refactor: move NuclearPowerSystem, SolarPowerSystem to new powersystem.py
