+++ gmatpyplus/hardware/__init__.py
@@ -2,4 +2,5 @@ from .fieldofview import *
 from .fueltank import *
 from .hardware import *
 from .imager import *
+from .powersystem import *
 from .thruster import *
